deduplicate: extract shared logic to common modules
- Extract verify_contract to contracts/common/verify.rs (was duplicated
  in sync.rs and async.rs)
- Move StreamDecoder impls to common/stream_decoders.rs for orders,
  news, and scanner (were duplicated between sync.rs and async.rs with
  cfg gates)
