+++ src/contracts/common/mod.rs
@@ -3,6 +3,7 @@
 pub(crate) mod decoders;
 pub(crate) mod encoders;
 pub(crate) mod stream_decoders;
+pub(crate) mod verify;
 
 #[cfg(test)]
 pub(crate) mod test_tables;
