+++ src/contracts/sync.rs
@@ -1,4 +1,4 @@
-use super::common::{decoders, encoders};
+use super::common::{decoders, encoders, verify};
 use super::*;
 use crate::client::blocking::{ClientRequestBuilders, Subscription};
 use crate::common::request_helpers;
@@ -9,26 +9,6 @@ use crate::{client::sync::Client, Error};
 use log::{error, info};
 
 impl Client {
-    fn verify_contract(&self, contract: &Contract) -> Result<(), Error> {
-        if !contract.security_id_type.is_empty() || !contract.security_id.is_empty() {
-            check_version(self.server_version, Features::SEC_ID_TYPE)?
-        }
-
-        if !contract.trading_class.is_empty() {
-            check_version(self.server_version, Features::TRADING_CLASS)?
-        }
-
-        if !contract.primary_exchange.is_empty() {
-            check_version(self.server_version, Features::LINKING)?
-        }
-
-        if !contract.issuer_id.is_empty() {
-            check_version(self.server_version, Features::BOND_ISSUERID)?
-        }
-
-        Ok(())
-    }
-
     /// Requests contract information.
     ///
     /// Provides all the contracts matching the contract provided. It can also be used to retrieve complete options and futures chains. Though it is now (in API version > 9.72.12) advised to use [Client::option_chain] for that purpose.
@@ -51,7 +31,7 @@ impl Client {
     /// }
     /// ```
     pub fn contract_details(&self, contract: &Contract) -> Result<Vec<ContractDetails>, Error> {
-        self.verify_contract(contract)?;
+        verify::verify_contract(self.server_version, contract)?;
 
         let builder = self.request();
         let request_id = builder.request_id();
@@ -439,7 +419,7 @@ mod tests {
             });
 
             let client = Client::stubbed(message_bus, test_case.server_version);
-            let result = client.verify_contract(&test_case.contract);
+            let result = verify::verify_contract(client.server_version, &test_case.contract);
 
             if test_case.should_error {
                 assert!(result.is_err(), "Test '{}' should have failed", test_case.name);
