+++ src/orders/common/mod.rs
@@ -2,4 +2,5 @@ pub(super) mod decoders;
 pub(crate) mod encoders;
 /// Helpers for constructing commonly used order templates.
 pub mod order_builder;
+pub(crate) mod stream_decoders;
 pub(super) mod verify;
