+++ src/scanner/common/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod decoders;
 pub(crate) mod encoders;
+pub(crate) mod stream_decoders;
